requirements.txt
google-genai==1.50.1
